requirements.txt
# === Core FastAPI + servidor ===
fastapi==0.110.2
uvicorn[standard]==0.30.1

# === LlamaIndex 0.10.x (compatibles entre sí) ===
llama-index-core==0.10.64
llama-index-llms-ollama==0.2.2
llama-index-embeddings-huggingface==0.2.2
llama-index-vector-stores-chroma==0.1.8
llama-index-readers-file==0.1.33

# === Vector DB ===
chromadb==0.5.5

# === NLP / embeddings ===
sentence-transformers==2.7.0
transformers==4.44.2
torch==2.4.0+cu121          # ✅ versión CUDA (GPU)
nltk==3.8.1

# === Utilidades varias ===
python-dotenv==1.0.1
requests==2.32.3

# === (Opcional, para PDFs) ===
pypdf==4.2.0
unstructured==0.15.7
orjson==3.10.7

# === Fuente oficial de ruedas CUDA para PyTorch ===
--extra-index-url https://download.pytorch.org/whl/cu121